refactor: move history.ts and historySearch.ts to history/
